modified:   app/main.cpp 	modified:   include/DataClassSet.h 	modified:   include/Matrix.h 	modified:   include/RTK.h 	modified:   include/ReadBinary.h 	modified:   include/SPP.h 	modified:   include/mathf.h 	modified:   include/output.h 	modified:   include/param.h 	modified:   src/LSQ.cpp 	modified:   src/Matrix.cpp 	modified:   src/RTK.cpp 	modified:   src/RTKscript.cpp 	modified:   src/ReConfig.cpp 	modified:   src/ReadBinary.cpp 	modified:   src/SPP.cpp 	modified:   src/SPPscript.cpp 	modified:   src/TimeCordination.cpp 	modified:   src/basCal.cpp 	modified:   src/output.cpp

+++ src/ReConfig.cpp
@@ -168,7 +168,7 @@ namespace rr
 		{
 			return default_value;
 		}
-		return stoi(it_item->second.c_str());
+		return atoi(it_item->second.c_str());
 	}
 
 	float RrConfig::ReadFloat(const char* section, const char* item, const float& default_value)
@@ -189,7 +189,7 @@ namespace rr
 		{
 			return default_value;
 		}
-		return stof(it_item->second.c_str());
+		return atof(it_item->second.c_str());
 	}
 	double RrConfig::ReadDouble(const char* section, const char* item, const double& default_value)
 {
